Removed Restriction of uploading notes exported by Evernote
Removed the comment in upload.html that stated uploads where for notes
that did not contain the [CDATA[ stuff.

Added Regex search in the upload function to do a search for the CDATA
header and footer info, then extract the content in the middle and use
that to upload the note. Also flash the return value of new note now so
that if there is an error it's caught.